+++ tests/src/bytecode/tests/functions/mod.rs
@@ -1,10 +1,6 @@
 mod compute_bytecode_root;
-mod compute_bytecode_root_with_configurables;
 mod compute_predicate_address;
-mod compute_predicate_address_with_configurables;
 mod predicate_address_from_root;
 mod swap_configurables;
 mod verify_contract_bytecode;
-mod verify_contract_bytecode_with_configurables;
 mod verify_predicate_address;
-mod verify_predicate_address_with_configurables;
