+++ tests/src/bytecode/tests/functions/swap_configurables.rs
@@ -142,7 +142,7 @@ mod failure {
     #[tokio::test]
     #[should_panic]
     async fn able_to_buffer_overflow() {
-        let (test_contract_instance, wallet) = test_contract_instance().await;
+        let (test_contract_instance, _wallet) = test_contract_instance().await;
         let (_contract_offset, _predicate_offset, config_value) = defaults();
 
         // Get the bytecode for the contract
@@ -152,7 +152,7 @@ mod failure {
         let my_configurables = build_simple_configurables(file_bytecode.len() as u64, config_value);
 
         // Call the contract to swap the configurables
-        let result_bytecode = swap_configurables(
+        let _result_bytecode = swap_configurables(
             &test_contract_instance,
             file_bytecode.clone(),
             my_configurables.clone(),
